Move getColumnKey into a util
