+++ src/modules/staffAssignment/components/MyClients.tsx
@@ -3,7 +3,8 @@ import Typography from '@mui/material/Typography';
 import { memoize } from 'lodash-es';
 import React from 'react';
 import { renderCellContents } from '@/components/elements/table/GenericTable';
-import { ColumnDef, getColumnKey } from '@/components/elements/table/types';
+import { ColumnDef } from '@/components/elements/table/types';
+import { getColumnKey } from '@/components/elements/table/util';
 import useAuth from '@/modules/auth/hooks/useAuth';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import { HouseholdStatus } from '@/modules/hmis/components/EnrollmentStatus';
